Rebased, organized tests into respective specs
