+++ apps/web/e2e/hosted-channels-providers.pw.ts
@@ -1,5 +1,6 @@
 import { expect, test } from "@playwright/test";
 import { CLOUD_CHANNEL_ID, collectBrowserErrors, stubCloudApi } from "./hosted-fixtures";
+import { expectNonZeroBox } from "./hosted-stub-api";
 
 test.beforeEach(async ({ page }) => {
 	await stubCloudApi(page);
@@ -39,4 +40,25 @@ test("AI providers BYOK flow saves and renders a custom provider", async ({ page
 	await expect(page.getByText("OpenAI", { exact: true }).first()).toBeVisible();
 	await expect(page.getByText("Vault key").first()).toBeVisible();
 	expect(errors, `providers flow: ${errors.join(" | ")}`).toEqual([]);
+});
+
+test("channels connect dialog opens without browser errors", async ({ page }) => {
+	const errors = collectBrowserErrors(page);
+	await page.goto("/channels");
+
+	const connect = page.getByRole("button", { name: /connect a bot/i }).first();
+	await expect(connect).toBeVisible();
+	await page.waitForTimeout(150);
+	expect(errors, `channels render: ${errors.join(" | ")}`).toEqual([]);
+
+	await expect(page.locator('[data-slot="tabs-list"]')).toHaveCount(0);
+	await expect(page.getByText("Your channels").first()).toBeVisible();
+	await expect(page.getByText("Shared bots").first()).toBeVisible();
+	await expectNonZeroBox(page.locator('[data-sidebar="separator"]').first(), "sidebar separator");
+
+	// Open the Base UI Dialog + interact with its provider picker.
+	await connect.click();
+	await expect(page.locator('[data-slot="dialog-content"]').first()).toBeVisible();
+	await page.waitForTimeout(150);
+	expect(errors, `connect dialog: ${errors.join(" | ")}`).toEqual([]);
 });
